memvid/memvid/requirements.txt
qrcode[pil]==8.2
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
sentence-transformers==4.1.0
numpy==1.26.4
tqdm==4.67.1
faiss-cpu==1.7.4
Pillow==11.2.1

# Optional dependencies for PDF support
PyPDF2==3.0.1
python-dotenv==1.1.0

# Optional imports for epub
beautifulsoup4
ebooklib

# Optional LLM provider imports
openai==1.82.0
google-generativeai==0.8.5
anthropic~=0.52.1
